Further testing to make sure that functionality of MyAnchor worked.
The problem was that NumberedPager was re-enabling MyAnchor under many conditions. I had to break these corner cases and put quite a few breakpoints to nail the behaviour. I can confirm the original behaviour (which I wasn't sure) of simply disabling the click handler, but not changing any styling (including the appearance of clickable).

In retrospect I should have just created a separate MyAnchor, unconnected to the behaviour of NumberedPager to test this... but such is hindsight.