Removed SimplePager from TopicPage.
Log: Busy following NumberedPager.setPageStart(index) for which the correct value of zero is passed when you click the link.

Down through the stack. By the time ForumMessageProvider.onRangeChanged is called, the index value is incorrect. So something
in the CellList mechanics is not liking that we set it to zero under some conditions.

Trying to follow further, but knowing exactly what is happening is difficult because there are a lot of unfamiliar constructs.