Forum: Fixed what seems to be a bug in the server starting pager for the 2nd page.
More testing required as it seems to still have bugs.
Also, the big problem, with the TopicPage messages not firing an onLoad event was because the server was missing the amount of items expected.
The Cell List never flagged that this was a problem (continue spinning, or throw an error), it just outputted the rows it had, stopped the spinner and never
fired onLoad.... not great communication.