RWA-326: Fix account linking issue on live
Task-URL: https://reflection.atlassian.net/browse/RWA-326

iTunes now gives a daily summary past 30 days, hence the old way of
checking for an exception does not work. This generated a NPE as the
verification code provided nulls for bucket name and path for saving the
sales file. This fix just returns a null of the bucket name or path are
null of empty and does not throw an exception.