Added ability to split data in DataAccountIngestorITunesConnect
Now when we get a product type identifier for universal apps, we use the
device column to identify the actual device the sales entry deals with
and update the type identifier to match.

EG: if we get a code of 1F (Universal Paid App download), we change it
to 1 or 1T if the device is iphone/ipod or ipad respectively.

The same with 3F and 7F (redownload and update)