Updated the account linking and unlinking processes to support multiples
Now when accounts are linked, we check for same vendor ids and add the
new user to all existing vendor id matched data accounts.

We also match the older accounts that match the username and vendorid.
On unlinking we removed the concept of an account owner, now just that
one user is unlinked from the data account.